Only delete the branch after merging if it is safe

+++ app/models/shipit/pull_request.rb
@@ -164,7 +164,9 @@ module Shipit
         merge_method: 'merge',
       )
       begin
-        Shipit.github_api.delete_branch(stack.github_repo_name, branch)
+        if Shipit.github_api.pull_requests(stack.github_repo_name, base: branch).empty?
+          Shipit.github_api.delete_branch(stack.github_repo_name, branch)
+        end
       rescue Octokit::UnprocessableEntity
         # branch was already deleted somehow
       end
